+++ src/cryptoadvance/specter/devices/cobo.py
@@ -3,8 +3,9 @@ import hashlib
 from .sd_card_device import SDCardDevice
 from hwilib.serializations import PSBT
 from binascii import a2b_base64
-from .. import bcur
+from ..util import bcur
 from .electrum import b43_encode
+from ..util.xpub import get_xpub_fingerprint
 
 
 class Cobo(SDCardDevice):
