+++ src/cryptoadvance/specter/user.py
@@ -1,7 +1,53 @@
-import os, shutil
+import os
+import shutil
+import hashlib
+import binascii
+import json
 from flask_login import UserMixin
-from .helpers import get_users_json, save_users_json
 from .specter_error import SpecterError
+from .helpers import fslock
+
+def hash_password(password):
+    """Hash a password for storing."""
+    salt = binascii.b2a_base64(hashlib.sha256(os.urandom(60)).digest()).strip()
+    pwdhash = binascii.b2a_base64(hashlib.pbkdf2_hmac('sha256', password.encode('utf-8'), salt, 10000)).strip().decode()
+    return { 'salt': salt.decode(), 'pwdhash': pwdhash }
+
+
+def verify_password(stored_password, provided_password):
+    """Verify a stored password against one provided by user"""
+    pwdhash = hashlib.pbkdf2_hmac('sha256', 
+                                  provided_password.encode('utf-8'), 
+                                  stored_password['salt'].encode(), 
+                                  10000)
+    return pwdhash == binascii.a2b_base64(stored_password['pwdhash'])
+
+def get_users_json(specter):
+    users = [
+        {
+            'id': 'admin',
+            'username': 'admin',
+            'password': hash_password('admin'),
+            'is_admin': True
+        }
+    ]
+
+    # if users.json file exists - load from it
+    if os.path.isfile(os.path.join(specter.data_folder, "users.json")):
+        with fslock:
+            with open(os.path.join(specter.data_folder, "users.json"), "r") as f:
+                users = json.loads(f.read())
+    # otherwise - create one and assign unique id
+    else:
+        save_users_json(specter, users)
+    return users
+
+
+def save_users_json(specter, users):
+    with fslock:
+        with open(os.path.join(specter.data_folder, 'users.json'), "w") as f:
+            f.write(json.dumps(users, indent=4))
+
 
 class User(UserMixin):
     def __init__(self, id, username, password, config, is_admin=False):
